static: record api.ruyisdk.cn/fe/dashboard
