+++ package.json
@@ -7,7 +7,7 @@
     "check:news-images": "node ./scripts/check-news-images.cjs",
     "check:news-images:strict": "node ./scripts/check-news-images.cjs --strict",
     "generate-github": "node ./scripts/generate-api-github.cjs",
-    "generate-latest-pm": "node ./scripts/generate-api-latest-pm.cjs",
+    "generate-ruyisdk-api": "node ./scripts/generate-api-ruyisdk-cn.cjs",
     "dev": "dotenv -e .env.development -- cross-env NODE_ENV=development docusaurus start",
     "start": "dotenv -e .env.production -- cross-env NODE_ENV=production docusaurus start",
     "build": "dotenv -e .env.production -- cross-env NODE_ENV=production docusaurus build",
